MedicalExpenseAnalysisTool/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
xgboost==1.4.2
pandas==1.3.0
